[Add] ExceptionHandlingMiddleware.cs

+++ PriceNegotiationAPI.Application/Product/Command/Delete/DeleteProductCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using PriceNegotiationAPI.Application.Abstraction;
+using PriceNegotiationAPI.Domain.Exceptions;
 using PriceNegotiationAPI.Domain.Repository;
 
 namespace PriceNegotiationAPI.Application.Product.Command.Delete;
@@ -13,6 +14,12 @@ internal class DeleteProductCommandHandler : ICommandHandler<DeleteProductComman
     }
     public async Task Handle(DeleteProductCommand request, CancellationToken cancellationToken)
     {
+        var productToDelete = await _productRepository.GetProductByIdAsync(request.ProductId, cancellationToken);
+        if (productToDelete is null)
+        {
+            throw new ProductNotFoundException($"Product with ID {request.ProductId} was not found.");
+        }
+        
         await _productRepository.DeleteProductAsync(request.ProductId, cancellationToken);
     }
 }
